fake-currency-python/requirements.txt
streamlit==1.36.0
opencv-python==4.10.0.84
numpy==1.26.4
pillow==10.3.0
